sidecar/requirements.txt
openai-whisper
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
